Connection_Service/requirements.txt
fastapi
grpcio
grpcio-tools
paho-mqtt
